dolphin/translator.py
import os
import pandas as pd

# TODO: DUMMY CODE FOR TRANSLATION
# THIS FUNCTION WILL BE REPLACED BY THE ACTUAL TRANSLATION LOGIC
# NOT SURE WHAT IS THE EXACT LOGIC FOR TRANSLATION FOR NOW, SO JUST RETURNING THE EXTRACTED TEXT AS IT IS. -- Alan

def translate_text(string_data):
    """
    Processes extracted text from an image to generate code based on predefined commands.

    :param string_data: Extracted text from an image.
    """

    file_name = "code_file.js"
    if os.path.exists(file_name):
        os.remove(file_name)

    forever_loop_default_if = ""
    forever_if_condition_input_1 = "true"
    forever_if_condition_input_2 = "true"
    forever_if_condition_input_3 = "true"
    forever_if_condition_input_4 = "true"
    forever_if_condition_input_5 = "true"
    forever_loop_elseif_cond_input = "true"
    forever_loop_elseif_default_if = ""
    forever_loop_else_default = ""
    and_or = ""
    code_channel = ""
    radio_received_if_condition_input = "HI"
    radio_received_loop_default_if = ""
    channel_no = 0
    loop_change_index = 100
    forever_else_if_change_index_jump = 100
    forever_else_change_index_jump = 100

    # Process commands from extracted text
    commands = string_data.split("\n")
    print(commands)
    df = pd.DataFrame({"command": commands, "full_command": 0, "code_command": 0})

    # Processing loop to handle different commands
    for index, row in df.iterrows():
        value = row['command']

    # Place the code for translation here
    # Dummy code for now

    return "Code generated successfully. Please check the code_file.js for the generated code."